gluon-radv-filterd: expire routers for redirect
For the redirect login implemented for parker, it is beneficial to keep
a list of seen routers and mark them as expired instead of removing them
from the global routers list. This way we don't enter a new redirect
entry for the same router when it shows up again after a mesh link
failure.

All foreach loops should be adjusted to correctly ignore expired
routers.

Signed-off-by: Rouven Czerwinski <rouven@czerwinskis.de>